Feat: Implementa o DTO e o Mapper de books no controller e service

+++ src/main/java/lest/dev/RecoBook/Book/service/BookService.java
@@ -1,5 +1,7 @@
 package lest.dev.RecoBook.Book.service;
 
+import lest.dev.RecoBook.Book.dto.BookDTO;
+import lest.dev.RecoBook.Book.mapper.BookMapper;
 import lest.dev.RecoBook.Book.model.Book;
 import lest.dev.RecoBook.Book.repository.BookRepository;
 import lest.dev.RecoBook.User.model.User;
@@ -8,50 +10,60 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class BookService {
 
     private final BookRepository bookRepository;
-    private final UserRepository userRepository;
+    private final BookMapper bookMapper;
 
-    public BookService(BookRepository bookRepository, UserRepository userRepository) {
+    public BookService(BookRepository bookRepository, BookMapper bookMapper) {
 
         this.bookRepository = bookRepository;
-        this.userRepository = userRepository;
+        this.bookMapper = bookMapper;
 
     }
 
-    public Book createBook(Book bookModel) {
-        return bookRepository.save(bookModel);
-
+    public BookDTO createBook(BookDTO bookDTO) {
+        Book book = bookMapper.map(bookDTO);
+        return bookMapper.map(bookRepository.save(book));
     }
 
-    public Book alterBook(Long id, Book bookModel) {
+    public BookDTO alterBook(Long id, BookDTO bookDTO) {
         Optional<Book> book = bookRepository.findById(id);
         if (book.isPresent()) {
-            bookModel.setId(id);
-            return bookRepository.save(bookModel);
+            Book bookAtt = bookMapper.map(bookDTO);
+            bookAtt.setId(id);
+            bookRepository.save(bookAtt);
+            return bookMapper.map(bookAtt);
         }
         return null;
     }
 
-    public void deleteBook(Long id) {
-
-        bookRepository.deleteById(id);
-
+    public boolean deleteBook(Long id) {
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()){
+            bookRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public List<Book> listBooks () {
-
-        return bookRepository.findAll();
-
+    public List<BookDTO> listBooks () {
+        List<Book> books = bookRepository.findAll();
+        return books.stream()
+                .map(bookMapper::map)
+                .collect(Collectors.toList());
     }
 
-    public Optional<Book> listByIdBook(Long id) {
-
-        return bookRepository.findById(id);
-
+    public BookDTO listByIdBook(Long id) {
+        Optional<Book> bookOptional = bookRepository.findById(id);
+        if (bookOptional.isPresent()) {
+            Book book = bookOptional.get();
+            return bookMapper.map(book);
+        }
+        return null;
     }
